+++ srcs/texture/cub3D_texture.c
@@ -6,7 +6,7 @@
 /*   By: jebouche <jebouche@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/14 20:10:04 by thuynguy          #+#    #+#             */
-/*   Updated: 2023/08/30 16:24:56 by jebouche         ###   ########.fr       */
+/*   Updated: 2023/08/31 18:02:37 by jebouche         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ void	load_texture(t_scene *scene, t_cubed *cubed)
  * line size and the projection height. The x is calculated by the direction 
  * of the ray and the position of the hit. 
 */
-static t_vector	get_texture_vec(t_vector *hit, int dir, float y_step, \
-float line_size)
+static t_vector	get_texture_vec(t_vector *hit, int dir, double y_step, \
+double line_size)
 {
 	t_vector	texture;
 
@@ -54,9 +54,9 @@ float line_size)
  * line height is greater than the projection height. Otherwise, it rounds the
  * line height to the nearest integer.
 */
-static float	clip_line_height(float to_clip)
+static double	clip_line_height(double to_clip)
 {
-	float	line_height;
+	double	line_height;
 
 	if (to_clip > PROJECTION_HEIGHT)
 		line_height = PROJECTION_HEIGHT;
@@ -79,9 +79,9 @@ static float	clip_line_height(float to_clip)
 void	draw_textured_walls(t_cubed *cubed, int project_x, t_ray_calc *ray, \
 int dir)
 {
-	float			line_height;
+	double			line_height;
 	unsigned int	pane_y_color[2];
-	float			y_step;
+	double			y_step;
 	t_vector		texture;
 	int				drawn_count;
